refactor/#5 : User.ts 파일 이름 수정
