+++ frontend/src/types/user.ts
